+++ np.Templating/lib/helpers.js
@@ -4,25 +4,3 @@
  * -----------------------------------------------------------------------------------------*/
 
 // @flow
-
-/**
- * Execute `fetch` calls with timeout
- * @author @codedungeon
- * @param {string} resource
- * @param {any} options
- * @returns {any}
- */
-export async function fetchWithTimeout(resource: string, options: any = {}): Promise<any> {
-  // default timeout 10 seconds
-  const { timeout = 10000 } = options
-
-  const controller = new AbortController()
-  const id = setTimeout(() => controller.abort(), timeout)
-  const response = await fetch(resource, {
-    ...options,
-    signal: controller.signal,
-  })
-  clearTimeout(id)
-
-  return response
-}