+++ np.Templating/lib/support/modules/quote.js
@@ -5,14 +5,9 @@
 
 // @flow
 
-// NOTE: Not using `@helpers/dev` here because it can't be resolved in jest tests
-//       this should serve as a strong reason to support module aliases such as `@helpers`
-//       as this is an ugly import
-import { fetchWithTimeout } from '../../../../helpers/dev'
-
 // $FlowFixMe
 export async function getDailyQuote(): Promise<string> {
-  const response = await fetchWithTimeout(`https://zenquotes.io/api/random`)
+  const response = await fetch(`https://zenquotes.io/api/random`, { timeout: 3000 })
   if (response) {
     //$FlowIgnore[incompatible-call]
     const quoteLines = JSON.parse(response)