+++ np.Templating/lib/support/modules/weather.js
@@ -5,16 +5,11 @@
 
 // @flow
 
-// NOTE: Not using `@helpers/dev` here because it can't be resolved in jest tests
-//       this should serve as a strong reason to support module aliases such as `@helpers`
-//       as this is an ugly import
-import { fetchWithTimeout } from '../../../../helpers/dev'
-
 export async function getWeather(): Promise<string> {
   try {
-    // return 'wttr.in unreachable'
     // $FlowFixMe
-    return await fetchWithTimeout('https://wttr.in?format=3')
+    let response: any = await await fetch(`https://wttr.in?format=3`, { timeout: 3000 })
+    return response ? response : '**An error occurred accessing weather service**'
   } catch (error) {
     return '**An error occurred accessing weather service**'
   }